+++ CHANGELOG.md
@@ -16,6 +16,16 @@ Possible log types:
 - [changed] Various small refactors (thanks sftse)
 - [changed] `Config::rich()` no longer includes decorations around `<em>` etc. - 
   use `Config::rich().do_decorate()` to get the old behaviour.
+- [fixed] Remove unnecessary empty lines at the start of lists (thanks russellbanks)
+- [added] New CSS support: `::before`/`::after` and `content: "string"`, which is now
+  used for simple decorations.  With CSS enabled, this allows for customising
+  the display of `<em>foo</em>` without writing a decorator.
+- [added] Add support for `<h5>` and `<h6>` (thanks noahbaculi)
+- [changed] Link footnotes are now configurable independently of the decorator, and on
+  by default for `config::plain()` but can be enabled or disabled with
+  `config.link_footnotes(true/false)`.  The footnote references (e.g. `[1]`) are added
+  in the main renderer, and the actual footnotes are written in a default implementation
+  of `TextDecorator::finalise()` so can be customised.
 
 ### 0.13.6